pyproject.toml
[project]
name = "narrative-game-base"
version = "0.1.0"
description = "A multi-agent narrative game system using LangGraph"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.0.10",
    "langchain-google-genai>=0.0.5",
    "pydantic>=2.0.0",
    "langchain-core",
    "python-dotenv>=1.2.1",
]

[tool.uv]
python-version = "3.11"
